+++ contracts/fastUpdates/implementation/FastUpdaters.sol
@@ -1,14 +1,10 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.18;
 
-import { ECPoint } from "../lib/Sortition.sol";
+import {ECPoint} from "../lib/Sortition.sol";
 
 contract FastUpdaters {
-    function numParticipants() public returns (uint) {
-        
-    }
+  function numParticipants() public returns (uint8) {}
 
-    function sortitionPublicKey(address provider) public returns (ECPoint memory) {
-
-    }
+  function sortitionPublicKey(address provider) public returns (ECPoint memory) {}
 }
